+++ pepper/rc/changelog.md
@@ -16,6 +16,7 @@
 - fix `` [aA]` `` was missing
 - added `log` as a `open` command option which is the same as old `scratch` and now `scratch` enables buffer history by default
 - added `file-backed` as a `open` command option which affects open and reopen commands
+- added `output` as a `open` command option which disables all buffer prorperties (suitable for process outputs)
 - added `to-lowercase` and `to-uppercase` commands
 - fix buffer views would not be removed when their clients closed
 - fix breakpoints locations on text edits